添加标签

+++ src/main/java/com/art1001/supply/controller/ShareController.java
@@ -123,21 +123,36 @@ public class ShareController {
         JSONObject jsonObject = new JSONObject();
         try {
             Share share = shareService.findById(shareId);
-            String[] tagIdArr = share.getTagIds().split(",");
-            if (CommonUtils.useList(tagIdArr, tagId)) { // 已经存在，移除
-                StringBuilder tagIds = new StringBuilder();
-                for (String tId : tagIdArr) {
-                    if (!tagId.equals(tId)) {
-                        tagIds.append(tId).append(",");
+            if (StringUtils.isNotEmpty(share.getTagIds())) {
+                String[] tagIdArr = share.getTagIds().split(",");
+                if (CommonUtils.useList(tagIdArr, tagId)) { // 已经存在，移除
+                    StringBuilder tagIds = new StringBuilder();
+                    for (String tId : tagIdArr) {
+                        if (!tagId.equals(tId)) {
+                            tagIds.append(tId).append(",");
+                        }
                     }
+                    if (StringUtils.isNotEmpty(tagIds)) {
+                        tagIds.deleteCharAt(tagIds.length() - 1);
+                    }
+                    shareService.deleteTag(shareId, tagIds.toString());
+                    jsonObject.put("result", 2);
+                    jsonObject.put("msg", "移除成功");
+                } else { // 不存在添加
+                    String tagIds = share.getTagIds();
+                    if (StringUtils.isNotEmpty(share.getTagIds())) {
+                        tagIds += "," + tagId;
+                    } else {
+                        tagIds = tagId;
+                    }
+
+                    shareService.deleteTag(shareId, tagIds);
+                    Tag tag = tagService.findById(Integer.valueOf(tagId));
+                    jsonObject.put("result", 1);
+                    jsonObject.put("data", tag);
+                    jsonObject.put("msg", "添加成功");
                 }
-                if (StringUtils.isNotEmpty(tagIds)) {
-                    tagIds.deleteCharAt(tagIds.length() - 1);
-                }
-                shareService.deleteTag(shareId, tagIds.toString());
-                jsonObject.put("result", 2);
-                jsonObject.put("msg", "移除成功");
-            } else { // 不存在添加
+            } else {
                 String tagIds = share.getTagIds();
                 if (StringUtils.isNotEmpty(share.getTagIds())) {
                     tagIds += "," + tagId;
@@ -151,6 +166,7 @@ public class ShareController {
                 jsonObject.put("data", tag);
                 jsonObject.put("msg", "添加成功");
             }
+
 
         } catch (Exception e) {
             log.error("添加标签异常, {}", e);
